--- go.mod
@@ -1,3 +1,0 @@
-module be101
-
-go 1.23.2
